Add http/1.x response module
