+++ src/info/http1/mod.rs
@@ -4,3 +4,4 @@
 //! and accompanying headers.
 
 pub mod header;
+pub mod response;
